backend/requirements.txt
Django==4.2
django-cleanup==9.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
pillow==11.1.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0